#55 refactor gamepads store, replace event listeners into store, delete hook events in store and replace interval into initialize and terminate functions, add index files for store and composables folders, update import paths
