+++ src/components/organisms/ODiagnostic.vue
@@ -1,7 +1,7 @@
 ﻿<script setup lang="ts">
 import { defineComponent } from 'vue';
 import { storeToRefs } from 'pinia';
-import { useGamepadsStore } from '@/store/useGamepadsStore';
+import { useGamepadsStore } from '@/store';
 import { MDiagnosticItem } from '@/components/molecules';
 
 defineComponent({
