+++ src/components/organisms/OManual.vue
@@ -1,7 +1,7 @@
 ﻿<script setup lang="ts">
 import { defineComponent, ref, onMounted, onUnmounted } from 'vue';
 import { storeToRefs } from 'pinia';
-import { useGamepadsStore } from '@/store/useGamepadsStore';
+import { useGamepadsStore } from '@/store';
 
 import type { TVibrator } from '@/models';
